+++ src/uniq-by.js
@@ -1,17 +1,14 @@
 // @flow
 
 /**
- * Uniq an array by a string key of objects in array
+ * Unique an array by a string key of objects in array
  * Returns only the values of the passed key
  * @example
  * const a = [ { foo: 'foo' }, { foo: 'bar' }, { foo: 'foo' } ]
- * uniqBy('foo', a) // => [ 'foo', 'bar' ]
+ * uniqBy('foo', a) // => [ { foo: 'foo' }, { foo: 'bar' } ]
  */
 
-// const uniqBy = <O: {+[string]: mixed}, K: $Keys<O>>(key: K) => (arr: O[]): $ElementType<O, K>[] =>
-//     [ ...new Set(arr.map((a) => a[key])) ]
-
-const uniqBy = <O: {+[string]: mixed}, K: $Keys<O>>(key: K, arr: O[]): $ElementType<O, K>[] =>
-  [ ...new Set(arr.map((a) => a[key])) ]
+// $FlowFixMe
+const uniqBy = (el: string, xs: Array<{+[string]: mixed}>) => xs.filter((e: mixed, i: number) => xs.findIndex((a: mixed) => a[el] === e[el]) === i)
 
 export default uniqBy