+++ src/uniq-by.test.js
@@ -5,7 +5,7 @@ import uniqBy from './uniq-by'
 
 test('uniqBy', (t): void => {
   const a = [ { foo: 'bar' }, { foo: 'bar' }, { foo: 'foo' } ]
-  const e = [ 'bar', 'foo' ]
+  const e = [ { foo: 'bar' }, { foo: 'foo' } ]
   t.deepEqual(uniqBy('foo', a), e, 'works')
   t.end()
 })