+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * `8.0.0` &mdash; **2018-08-03** **breaking**
+  * Change `uniqBy` to work as expected. See docs.
   * Remove all functional-style utils. I recommend using Ramda or Lodash for FP,
     since those libraries are better tested and maintained. Removed functions:
     * all.js